Resolve import cycle error
