+++ package.json
@@ -96,6 +96,9 @@
     "@nestjs/cli": "^8.1.3",
     "@nestjs/schematics": "^8.0.7",
     "@nestjs/testing": "^8.3.1",
+    "@swc/cli": "^0.1.62",
+    "@swc/core": "^1.3.42",
+    "@swc/jest": "^0.2.24",
     "@types/cache-manager": "^3.4.3",
     "@types/express": "^4.17.13",
     "@types/graphql": "^14.5.0",
@@ -120,9 +123,11 @@
     "husky": "^7.0.4",
     "jest": "^27.3.0",
     "lint-staged": "^12.1.3",
+    "nodemon": "^2.0.22",
     "prettier": "^2.5.1",
     "source-map-support": "^0.5.20",
     "supertest": "^6.2.2",
+    "swc": "^1.0.11",
     "ts-jest": "^27.0.7",
     "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
@@ -138,7 +143,7 @@
     "rootDir": "src",
     "testRegex": ".*\\.spec\\.ts$",
     "transform": {
-      "^.+\\.(t|j)s$": "ts-jest"
+      "^.+\\.(t|j)sx?$": "@swc/jest"
     },
     "collectCoverageFrom": [
       "**/*.(t|j)s"
