+++ src/App/pinJSONAndImage/pin.resolver.spec.ts
@@ -11,7 +11,7 @@ import { HttpModule } from '@nestjs/axios'
 import PinService from './pin.service'
 import PinResolver from './pin.resolver'
 import IPFSValidatorService from '../../Indexer/Validator/validator.service'
-import PinJSONErrorWebhook from './webhookHandler/pinJSONErrorWebhook'
+// import PinJSONErrorWebhook from './webhookHandler/pinJSONErrorWebhook'
 import MetadataChangesService from '../../Indexer/Store/metadataChanges.service'
 import WebhookHandler from '../utils/discordWebhookHandler'
 import ActivityService from '../Activities/activity.service'
@@ -40,7 +40,7 @@ describe('PinResolver', () => {
             secret: '',
           },
         },
-        PinJSONErrorWebhook,
+        // PinJSONErrorWebhook,
         WebhookHandler,
         MetadataChangesService,
         ActivityService,
